fixed zip function

+++ src/frame.hpp
@@ -164,7 +164,7 @@ private:
 	int frame_time_;
 	bool reverse_frame_;
 	bool play_backwards_;
-	int scale_;
+	float scale_;
 	int pad_;
 	int rotate_;
 	int blur_;
